event operations tests
